+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Change: Extract filename handling out from `admin_menu` and into new `get_filename()`
 * Change: Extract logic to determine if CSV button should be shown out from `admin_menu` and into new `should_show_csv_button()`
 * Change: Extract logic to determine if email addresses should be listed on the admin page out from `admin_menu` and into new `should_show_email_addresses()`
+* New: Add `get_plugin_basename()` to retrieve the plugin's basename
 * Unit tests:
     * New: Add test to verify plugin hooks `plugins_loaded` action to initialize itself
     * New: Add tests for filters: `c2c_commenter_emails_filename`, `c2c_commenter_emails_show_csv_button`, `c2c_commenter_emails_show_emails`